+++ CapstoneProject_FPT_Summer2024/Infrastructures/Repository/SchoolYearRepository.cs
@@ -38,14 +38,14 @@ namespace Infrastructures.Repository
                 .ToListAsync();
         }
 
-        public async Task<SchoolYear> GetYearBySchoolYearId(int schoolId, short year)
+        public async Task<SchoolYear> GetYearBySchoolYearId(int schoolId, int year)
         {
             return _context.SchoolYears
                 .Include(c => c.School)
                 .FirstOrDefault(s => s.SchoolId == schoolId && s.Year == year);
         }
 
-        public async Task<SchoolYear> GetSchoolYearBySchoolIdAndYear(int schoolId, short year)
+        public async Task<SchoolYear> GetSchoolYearBySchoolIdAndYear(int schoolId, int year)
         {
             return await _context.SchoolYears
                 .Include(c => c.School)
@@ -53,7 +53,7 @@ namespace Infrastructures.Repository
         }
 
         // get school year which status is ONGOING by school id and year
-        public async Task<SchoolYear> GetOngoingSchoolYearBySchoolIdAndYear(int schoolId, short year)
+        public async Task<SchoolYear> GetOngoingSchoolYearBySchoolIdAndYear(int schoolId, int year)
         {
             return await _context.SchoolYears
                 .Include(c => c.School)
